requirements.txt
requests==2.31.0
python-dotenv==1.0.0
selenium==4.18.1
webdriver-manager==4.0.1
beautifulsoup4==4.12.3
schedule==1.2.1
python-telegram-bot==20.8
aiohttp==3.9.3 